+++ nautilus_trader/model/orders/base.pyx
@@ -44,7 +44,6 @@ from nautilus_trader.model.events cimport OrderEvent
 from nautilus_trader.model.events cimport OrderExpired
 from nautilus_trader.model.events cimport OrderFilled
 from nautilus_trader.model.events cimport OrderInitialized
-from nautilus_trader.model.events cimport OrderInvalid
 from nautilus_trader.model.events cimport OrderPendingCancel
 from nautilus_trader.model.events cimport OrderPendingReplace
 from nautilus_trader.model.events cimport OrderRejected
@@ -62,7 +61,6 @@ from nautilus_trader.model.objects cimport Quantity
 
 # State being used as trigger
 cdef dict _ORDER_STATE_TABLE = {
-    (OrderState.INITIALIZED, OrderState.INVALID): OrderState.INVALID,
     (OrderState.INITIALIZED, OrderState.DENIED): OrderState.DENIED,
     (OrderState.INITIALIZED, OrderState.SUBMITTED): OrderState.SUBMITTED,
     (OrderState.SUBMITTED, OrderState.REJECTED): OrderState.REJECTED,
@@ -234,8 +232,7 @@ cdef class Order:
 
     cdef bint is_completed_c(self) except *:
         return (
-            self._fsm.state == OrderState.INVALID
-            or self._fsm.state == OrderState.DENIED
+            self._fsm.state == OrderState.DENIED
             or self._fsm.state == OrderState.REJECTED
             or self._fsm.state == OrderState.CANCELED
             or self._fsm.state == OrderState.EXPIRED
@@ -547,10 +544,7 @@ cdef class Order:
             Condition.equal(event.venue_order_id, self.venue_order_id, "event.venue_order_id", "self.venue_order_id")
 
         # Handle event (FSM can raise InvalidStateTrigger)
-        if isinstance(event, OrderInvalid):
-            self._fsm.trigger(OrderState.INVALID)
-            self._invalid(event)
-        elif isinstance(event, OrderDenied):
+        if isinstance(event, OrderDenied):
             self._fsm.trigger(OrderState.DENIED)
             self._denied(event)
         elif isinstance(event, OrderSubmitted):
@@ -603,9 +597,6 @@ cdef class Order:
 
         # Update events last as FSM may raise InvalidStateTrigger
         self._events.append(event)
-
-    cdef void _invalid(self, OrderInvalid event) except *:
-        pass  # Do nothing else
 
     cdef void _denied(self, OrderDenied event) except *:
         pass  # Do nothing else